firmware + common: Simplify the sampler design (prototype)
This commit reduces the width of the counter value to 6 bits instead of
7, leaving the 2 higher bits available to encode the state of the INDEX
and RDATA lines.
Each 8 bit opcode now encodes the state of both the INDEX and RDATA
lines, as well as the time since the last event.
Any combination of state change is always processed by the sampler in 2
clock cycles.

This simple change allows to save 2 bits on the USB for every RDATA and
INDEX pulse compared to the previous format.
On the other hand, it doubles the bandwidth usage in the situation when
no RDATA or INDEX event occur (i.e. counter reaches the maximum value),
which is a negligible issue since bandwidth usage is at its minimum.

Also, it removes the need to have data crunched by the CPU, possibly
allowing the USB block to source its bulk EP data directly from the
sampler buffers via DMA, avoiding copy by the CPU.

This commit is a proof of concept implementation of the new sampler
bitstream format.
It breaks a lot of things, most notably write support, and relies on
ugly hacks to work without having to rework most of the code, which
still relies internally on the old bitstream format.
The reworked sampler currently only works when the sampleclock is 1/3
of the bus clock frequency.

List of changes:
- Reworked Sampler verilog, as explained above
- Turned the cruncher code into a very inefficient memcpy()
- Hijack uncruncher code to convert new bitstream format into old
  bitstream format, which all the client processing code expects
- Disable cruncher/uncruncher tests, as they would obviously fail
- Mess with the clocks in a vain hope to pass the static timing
  analysis
- Disable INDEX300 and INDEX360 pulse generation to pass said STA
- Expose the debug LED through a register accessible by software
- Updated those undecipherable project files

Those changes were tested by reading a 3.5" HD IBM formatted disk,
which was not readable previously (not enough USB bandwidth, and
sometimes a memory corruption bug which was happening when the cruncher
couldn't keep up with the sampler FIFO).